Add app directory ensuring
