+++ README.md
@@ -19,7 +19,7 @@ It's main use is to be a backend for [mpv-web-front](https://github.com/sarpt/mp
 
 - `allow-cors` - bool - (default: `false`) whether Cross Origin Requests should be allowed
 - `addr` - string - (default: `:3001`) address used to host the server
-- `app-dir` - string - (default: ` `) path on the file system which is used for application as a storage directory for unnamed playlists, cache, configs etc. When not provided, application tries to deduce default application directory path first by trying `mwa` directory in user directory, and if that fails for whatever reason, then it tries to use `mwa` directory in temp directory. If the directory does not exist, application tries to create the directory - if it fails at creating the directory, the application fails to start.
+- `app-dir` - string - (default: ` `) path on the file system which is used for application as a storage directory for unnamed playlists, cache, configs etc. When not provided, application tries to deduce default application directory path first by trying `.mwa` directory in user directory, and if that fails for whatever reason, then it tries to use `.mwa` directory in temp directory. If the directory does not exist, application tries to create the directory - if it fails at creating the directory, the application fails to start.
 - `dir` - []string - (default: current working directory) directories that should be scanned for media files. To specify more than one directory to be handled, multiple `--dir=<path>` arguments can be specified eg. `--dir=/path1 --dir=/path2`. The server will only handle paths provided by clients that start with one of the paths provided to `dir`. When not provided, current working directory for the process will be used to scan for media files. Recursive scan can be enabled with `--dir-recursive`. Watching for the changes to the provided directories can be enabled with `--watch-dir`.
 - `dir-recursive` - bool - directories provided to `--dir` (or working directory when `--dir` is not provided) will be checked recursively.
 - `mpv-socket-path` - string - (default: `/tmp/mpvsocket`) path to socket file used by MPV instance
